==> R/02_NARA_S_EDA.Rmd <==
---
title: "NARA_S_DataExploration"
author: "Toon Van Daele"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(usdm)
library(tidyverse)
library(pander)
source("11_NARA_S_Functions.R")
```

```{r load_data}
df_plant <- readRDS("../data/data-in/df_plant.RDS")
df_expl <- readRDS("../data/data-in/df_expl.RDS")
df_ecoreg <- readRDS("../data/data-in/df_ecoreg.RDS")
df_redlist <- readRDS("../data/data-in/df_redlist.RDS")
```

# Data exploration of plant data

## Total number of plant species

```{r nb_species}
nb_spec <- df_plant %>%
  select(Taxoncode) %>%
  n_distinct()
```

`r nb_spec`

## Total number of distinct utmID grids with at least 100 plant species

```{r nb_utmgrids}
nb_utm <- df_plant %>%
  dplyr::select(utmID) %>%
  n_distinct()
```

`r nb_utm`

## List of species codes and full names
```{r speccodes}
species <- df_plant %>%
  select(-utmID) %>%
  distinct()
head(species)
```

# Selection of grids with at least 70% of 1x1km

(Except for grids along the coast line and along river Meuse)

```{r selectutm}
pgrid <- df_expl %>%
  filter(scen == "Current" & varname == "Opp_m2") %>%
  left_join(y = df_ecoreg %>%
              filter(regio == "Ecoregio van de grindrivieren" |
                     regio == "Ecoregio van de kustduinen"),
            by = "utmID") %>%
  mutate(opp_ok = ifelse(value > 700000 | !is.na(regio) , 1, 0))

ggplot(pgrid, aes(x = X, y = Y, colour = factor(opp_ok))) +
  geom_point(size = 0.05) + coord_fixed()

utmok <- pgrid %>%
  filter(opp_ok == 1) %>%
  pull(utmID)

df_expl <- filter(df_expl, utmID %in% utmok)
df_plant <- filter(df_plant, utmID %in% utmok)
```

# Selection of plant species

```{r selecttaxon}
speclist <- df_plant %>%
  group_by(Taxoncode) %>%
  count() %>%
  filter(n >= 30 & n <= 0.80 * nb_utm) %>%
  pull(Taxoncode)

temp <- df_redlist %>%
  filter(!is.na(RLweging)) %>%
  pull(Soortcode)

speclist <- speclist[speclist %in% temp]

# for testing 
speclist <- sample(speclist, 200)

```

A total of `r length(speclist)` species.

# Ligging van de goed geobserveerde utm grids (minstens 100 soorten)

```{r joindata}
df_plant$present <- 1
spr_plant <- df_plant %>%
  filter(Taxoncode %in% speclist) %>%
  select(utmID, Taxoncode, present) %>%
  spread(key = Taxoncode, value = present) %>%
  tbl_df()

spr_plant <- left_join(x = spr_plant,
                       y = df_expl %>%
                            filter(scen == "Current") %>%
                            distinct(utmID, X, Y),
                       by = "utmID")

ggplot(spr_plant, aes(x = X, y = Y)) + geom_point(size = 1) + coord_fixed()
```

Deze locaties zijn niet goed ruimtelijke gebalanceerd. Een selectie op basis
van nearest neighbours met een grts bestand verminderd de sterke concentratie.

```{r balancedlocationsGRTS}
library(RANN)
grts <- readRDS("../data/grts_red.RDS")

grts <- grts %>%
  filter(Ranking < 30000)

closest <- nn2(spr_plant[, c("X", "Y")], grts[, c("x", "y")],
               k = 1, searchtype = "radius", radius = 50000)

closest <- data.frame(id = closest$nn.idx)

ids <- closest %>%
  filter(id > 0) %>%
  distinct() %>%
  pull()

spr_plant <- spr_plant %>%
  mutate(rownb = 1:n(),
  sel = rownb %in% ids)

ggplot(spr_plant, aes(x = X, y = Y, colour = sel)) +
  geom_point(size = 1) + coord_fixed()
  
spr_plant <- filter(spr_plant, sel == TRUE) %>%
  select(-sel, -rownb)

df_plant <- filter(df_plant, utmID %in% spr_plant$utmID)

ggplot(spr_plant, aes(x = X, y = Y)) + geom_point(size = 1) + coord_fixed()
```

# Data exploration of environmental (explanatory) variables

## Land use

```{r lu_summary}
df_expl %>%
  filter(vargroup == "lu" & scen == "Current") %>%
  group_by(varname) %>%
  select(-X, -Y, -scen, -vargroup) %>%
  spread(key = varname, value = value, ) %>%
  summary()
```

```{r lu_histogram}
df_expl %>%
  filter(vargroup == "lu" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram() +
  facet_wrap(~varname, scales = "free_y")
```

slikschor en duinen komen bijna niet voor

utmID met slikschor > 0:
```{r slikschor}
df_expl %>%
  filter(varname == "slikschor" & scen == "Current" & value > 0) %>%
  select(utmID)
```

Dit is weinig informatief en geeft gegarandeerd problemen bij het modellering.
Slikschor wordt verwijderd uit de lijst met verklarende variabelen

```{r dropslikschor}
df_expl <- filter(df_expl, !varname == "slikschor")
```

```{r duinen}
nb_duinen <- df_expl %>%
  filter(scen == "Current" & varname == "duinen" & value > 0) %>%
  dplyr::n_distinct()
```

Duinen komt op 'r nb_duinen' locaties voor.

```{r moeras}
df_expl %>%
  filter(varname == "moeras" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram(bins = 100)
```

```{r lu_pairs}
set.seed(5)
df_expl %>%
  filter(vargroup == "lu" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen, -vargroup) %>%
  sample_n(200) %>%
  pairs(pch = 20, cex = 0.7)
```

```{r lu_cor, results='asis'}
df_expl %>%
  filter(vargroup == "lu" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen, -vargroup) %>%
  cor() %>%
  round(2) %>%
  pandoc.table()
```

"Stadsgroen", "Wegen" and "bebouwing" zijn vrij sterk gecorreleerd

```{r selectlu}
set.seed(5)
varsel <- c("stadsgroen", "wegen", "bebouwing")
df_expl %>%
  filter(varname %in% varsel & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen, -vargroup) %>%
  sample_n(500) %>%
  pairs(pch = 20, cex = 0.7)
```

We kiezen slechts één, namelijk "bebouwing"

```{r droplu}
df_expl <- filter(df_expl, !varname == "stadsgroen" & !varname == "wegen")
```

## Soil

```{r soil_summary}
df_expl %>%
  filter(vargroup == "soil" & scen == "Current") %>%
  spread(key = varname, value = value, ) %>%
  select(-utmID, -X, -Y, -scen, -vargroup) %>%
  summary() %>%
  pandoc.table()
```

```{r soil_histogram}
df_expl %>%
  filter(vargroup == "soil" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram() +
  facet_wrap(~varname, scales = "free_y")
```

Textuurmergkrijt zijn bijna allemaal 0

```{r soilmergkrijt}
df_expl %>%
  filter(varname == "textmergkrijt" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram()

nb_mergkrijt <- df_expl %>%
  filter(varname == "textmergkrijt" & scen == "Current" & value > 0) %>%
  n_distinct()
```

Slechts 'r nb_mergkrijt' grids met opp > 0

We laten textmergkrijt weg uit de selectie van variabelen

```{r dropmergkrijt}
df_expl <- filter(df_expl, !varname == "textmergkrijt")
```

```{r soil_pairs}
df_expl %>%
  filter(vargroup == "soil" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -vargroup, -scen) %>%
  sample_n(500) %>%
  pairs()
```

Zand, leem en klei sommeren tot 100%. De textuur wijzigt niet bij een
veranderend landgebruik. We kunnen daaarom de twee principale componenten
van de textuurvariabelen "zand", "leem" en "klei" nemen.
De variabele "textveen" wordt apart behouden.

```{r PCAcond}
soiltext <- c("textzand", "textleem", "textklei")
df_text <- df_expl %>%
  filter(varname %in% soiltext & scen == "Current") %>%
  spread(key = varname, value = value)

pctext <- prcomp(df_text[, soiltext], center = TRUE, scale. = TRUE)
summary(pctext)
```

We voegen de PCA loads voor PC1 and PC2 toe aan df_expl
Deze procedure gaat ervan uit dat de texttuur voor elk scenario hetzelfde is!

```{r PCAloads}
pcloads <- pctext$x %>%
  as.data.frame() %>%
  bind_cols(df_text) %>%
  rename(textpc1 = PC1, textpc2 = PC2) %>%
  select(-one_of(soiltext), -PC3, -scen) %>%
  gather(key = varname, value = value, -utmID, -X, -Y, -vargroup)

scens <- data.frame(scen = unique(df_expl$scen), stringsAsFactors = FALSE)
pcloads <- merge(pcloads, scens)
  
df_expl <- bind_rows(df_expl, pcloads)

df_expl %>%
  filter(vargroup == "soil" & scen == "Current" & !varname == "textveen") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -vargroup, -scen) %>%
  sample_n(500) %>%
  pairs()

df_expl <- filter(df_expl, !varname %in% c("textleem", "textklei", "textzand"))
```

(Deze benadering zou vervangen kunnen worden door een variabele "bulk density"
of iets dergelijks.)

## Site conditions

```{r condvarsummary, results='asis'}
df_expl %>%
  filter(vargroup == "cond" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen, -vargroup) %>%
  summary() %>%
  pandoc.table()
```

Er zijn NA's voor potnatzuur en potnattrofie. Waar zijn die gelegen?

```{r potnatzuur}
df_expl %>%
  filter(varname == "potnatzuur" & scen == "Current" & is.na(value)) %>%
  ggplot(aes(x = X, y = Y, colour = value)) + geom_point(size = 0.1) +
  coord_fixed()
```

Dit is in de haven van Zeebrugge. Deze utm locaties worden verwijderd uit
de analyse.

```{r droputm_NA}
utmna <- df_expl %>%
  filter(varname == "potnatzuur" & scen == "Current" & is.na(value)) %>%
  distinct(utmID) %>%
  pull(utmID)
df_expl <- filter(df_expl, !utmID %in% utmna)
```

```{r condvarhist}
df_expl %>%
  filter(vargroup == "cond" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram() +
  facet_wrap(~varname, scales = "free")
```

potnatzilt komt niet zoveel voor. Opp bijna overal = 0

```{r condpotnatzilt}
df_expl %>%
  filter(varname == "potnatzilt" & scen == "Current") %>%
  ggplot(aes(x = X, y = Y, colour = value)) + geom_point(size = 1) +
  coord_fixed()
```

Allicht voor bepaalde plantensoorten erg bepalend. Toch maar behouden.

```{r condvarpairs}
df_expl %>%
  filter(vargroup == "cond" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -vargroup, -scen) %>%
  sample_n(500) %>%
  pairs()
```

```{r condvarcorr, results='asis'}
df_expl %>%
  filter(vargroup == "cond" & scen == "Current") %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -vargroup, -scen) %>%
  cor() %>%
  round(2) %>%
  pandoc.table()
```

Een vrij sterke relatie tussen zuurgraad en trofie.

Drainage heeft een merkwaardige knik in de verdeling.
Er zijn zeer weinig waarden tussen 1 en 2 en dan plots heel veel bij 2.
Nog uit te zoeken hoe dit komt.

```{r histodrainage}
df_expl %>%
  filter(varname == "drainage" & scen == "Current") %>%
  ggplot(aes(x = value)) + geom_histogram(bins = 100)
```

```{r plotdrainage}
df_expl %>%
  filter(varname == "drainage" & scen == "Current") %>%
  filter(value > 1 & value < 2.2) %>%
  ggplot(aes(x = X, y = Y, colour = value)) + geom_point(size = 0.2) +
  ggtitle("drainage < 2.2")
```

# Selectie van de verklarende variabelen

We onderscheiden verder twee groepen.
* Variabelen die wijzigen bij gewijzigd landgebruik -> alle landgebruiksklassen
* Variabelen die (bijna) niet wijzigen bij gewijzigd landgebruik ->
textuurvariabelen (twee principale componenten), textveen en potnatzuur

We gaan ervan ui dat de standplaatsvariabelen "potnattrofie" en "drainage"
gebonden zijn aan het landgebruik. Voor drainage lijkt dit op het eerste
zicht niet onmiddellijk het geval. We behouden drainage en beschouwen die
constant over de verschillende scenarios. Potnattrofie
wordt niet verder gebruikt in de analyse.

```{r trofieenlandgebruik}
df_expl %>%
  select(-vargroup) %>%
  filter(scen == "Current") %>%
  filter(varname %in% c("heide", "loofbos", "akker", "potnattrofie")) %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen) %>%
  sample_n(500) %>%
  pairs()
```

```{r drainagelandgebruik}
df_expl %>%
  select(-vargroup) %>%
  filter(scen == "Current") %>%
  filter(varname %in% c("heide", "loofbos", "akker", "drainage")) %>%
  spread(key = varname, value = value) %>%
  select(-utmID, -X, -Y, -scen) %>%
  sample_n(500) %>%
  pairs()
```

```{r droppotnattrofie}
df_expl <-  filter(df_expl, !varname == "potnattrofie")
```

# Selectie van variabelen per soort

```{r selectvars}
species <- spr_plant %>%
  select(-utmID, -X, -Y) %>%
  colnames()

gvar <- map_dfr(species, getexplvar,
             spr_plant = spr_plant,
             df_expl = df_expl)

gvar %>%
  group_by(varname) %>%
  count() %>%
  pandoc.table()

# Filter op basis minimaal 5 distincte waarden

explspec <- gvar %>%
 filter(distinct > 5)

explspec %>%
  group_by(varname) %>%
  count() %>%
  pandoc.table()
```


# Normalize explanatory variables

! expl var for new scenarios require the same scale parameters

Geen normalisatie meer.

# Future scenarios - "kijkrichtingen"

```{r future_scenarios, results='asis'}
df_sum <- df_expl %>%
  group_by(scen, varname) %>%
  summarise(sum = sum(value)) %>%
  spread(key = varname, value = sum) %>%
  t %>%
  pandoc.table()
```

# Overal species richness based on presence data

```{r species_richness}
speccode <- df_plant %>%
  select(Taxoncode, NaamNederlands) %>%
  distinct()
 
dftemp <- spr_plant
dftemp[is.na(dftemp)] <- 0
pdata <- dftemp %>%
     select(utmID, X, Y, speclist) %>%
    gather(key = sp_n, value = bins, -utmID, -X, -Y) %>%
    left_join(speccode, by = c("sp_n" = "Taxoncode")) %>%
    group_by(utmID) %>%
    summarise(specrich = sum(bins)) %>%
  left_join(dftemp %>%
              select(utmID, X, Y),
            by = "utmID")

(p <- ggplot(pdata, aes(x = X, y = Y)) +
  geom_point(aes(colour = specrich), size = 1) +
  scale_colour_gradientn(colours = rainbow(3)) +
  coord_fixed())

ggsave("../data/data-in/specrich.png", p, dpi = 150)
```

## Save plots with calibration data (presences only)

```{r species_plot, eval=FALSE, include=FALSE}
mypath <- "../data/data-in/presences/"
dir.create(mypath, showWarnings = FALSE)

speccode <- df_plant %>%
  select(Taxoncode, NaamNederlands) %>%
  distinct()
 
dftemp <- spr_plant
dftemp[is.na(dftemp)] <- 0
dftemp %>%
     select(utmID, X, Y, speclist) %>%
    gather(key = sp_n, value = bins, -utmID, -X, -Y) %>%
    left_join(speccode, by = c("sp_n" = "Taxoncode")) %>%
    split(.$sp_n) %>%
    map(., plotbinaries, mypath, c("sp_n", "NaamNederlands"))
```

plots with presence data saved in 'r mypath'

```{r savedata}
saveRDS(speclist, "../data/data-in/speclist_eda.RDS")
saveRDS(explspec, "../data/data-in/explspec_eda.RDS")
saveRDS(df_expl, "../data/data-in/df_expl_eda.RDS")
saveRDS(df_plant, "../data/data-in/df_plant_eda.RDS")
```
