==> R/05_NARA_S_PostProcesModels.Rmd <==
---
title: "NARA_S_PostProcesModels"
author: "Toon Van Daele"
date: "`r Sys.Date()`"
output: html_document
params:
  modelname: "rlw"
  calcvarimp : FALSE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(pander)
library(RSQLite)
library(ggplot2)
source("11_NARA_S_Functions.R")
options(stringsAsFactors = FALSE)
modelname <- params$modelname #"allsp" 
calcvarimp <- params$calcvarimp
```

Postprocessing of the results of the model run: **`r modelname`**

Calculation of variable importance: `r calcvarimp`

```{r load data}
path <- paste0("../data/data-out/", modelname, "/")
db <- dbConnect(SQLite(), dbname = paste0(path, "speclist.sqlite"))
speclist <- dbReadTable(db, "speclist")
dbDisconnect(db)
```

path to data:  `r path`

### Evaluation data

```{r evaluationdata, message=FALSE}
myspecies <- speclist %>%
  filter(modelrun == "ok") %>%
  pull(species)
  
df_eval <- myspecies %>%
  map_dfr(., getevaluation, modelname)

df_eval$algo <- sub( "_RUN.*_PA*.", "", df_eval$Model.name) 

saveRDS(df_eval, file = paste0(path, "df_eval.RDS"))
```

# Calculate the mean ROC, TSS and Kappa evaluation scores

```{r calceval}
df_meval <- df_eval %>%
  dplyr::group_by(sp_n, algo, Eval.metric) %>%
  dplyr::summarise(mean = mean(Testing.data))
  
saveRDS(df_meval, file = paste0(path, "df_meval.RDS"))
```

# mean evaluation score

```{r meanAUC, results='asis'}
df_meval %>%
  unite(algo_metric, algo, Eval.metric, sep = "_") %>%
  spread(key = algo_metric, value = mean) %>%
  pandoc.table()
```

# Overal mean evaluation score

```{r overalmeanAUC, results='asis'}
df_meval %>%
  group_by(Eval.metric, algo) %>%
  summarise(meanScore = mean(mean, na.rm = TRUE)) %>%
  pandoc.table()
```

```{r speclistok}
speclist <- df_meval %>%
    filter(Eval.metric == "ROC", mean > 0.65) %>%
    pull(sp_n)
saveRDS(speclist, file = paste0(path, "speclist.RDS"))
```

### Create repsonse curves

```{r responscurves, eval=FALSE, include=FALSE}
dir.create(paste0(path, "/responsecurves/"), showWarnings = FALSE)

orig_wd <- getwd()
setwd("../data/models")

# Onduidelijk waarom. Maar de responscurve werkt niet via een functie.
# Heeft te maken met het laden van de modellen en modelnamen
# De code zou volgende lijn moeten bevatten:
# Toe te voegen: walk(speclist, CreateResponsePlot, modelname, path)

create_responseplot(speclist = myspecies,
                   modelname = modelname)

setwd(orig_wd)
```


### Retrieve variable importance

```{r vi, results='asis'}
if (calcvarimp == TRUE) {
  vi <- map_dfr(speclist$species, get_variable_importance, modelname)

  vimean <- vi %>%
    dplyr::filter(grepl("ROC", EMmodel)) %>%
    dplyr::group_by(sp_n, Varname) %>%
    dplyr::summarise(meanvi = mean(vi)) %>%
    tidyr::spread(key = Varname, value = meanvi) %>%
    pandoc.table()

  saveRDS(vi, file = paste0(path, "vi.RDS"))

  vi %>%
  filter(grepl("ROC", EMmodel)) %>%
  ggplot(aes(x = Varname, y = vi)) + geom_boxplot()
}  
```
