==> R/07_NARA_S_PostProcesProject.Rmd <==
---
title: "07_NARA_S_PostProcesProject"
author: "Toon Van Daele"
date: "`r Sys.Date()`s"
output: html_document
params:
  modelname: "rlw"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(pander)
source("11_NARA_S_Functions.R")
modelname <- params$modelname  #test
path <- paste0("../data/data-out/", modelname, "/")
```

Postprocessing of the projections of model: **`r modelname`**
Path to data: `r path`

```{r loaddata}
df_proj_in <- readRDS(paste0(path, "df_proj_in.RDS"))
speclist <- readRDS(paste0(path, "speclist.RDS"))
```

## Probabilities from biomod projections

```{r get probabilities}
projs <- expand.grid(sp_n = speclist,
                     scenario = unique(df_proj_in$scen),
                     stringsAsFactors = FALSE)
projs$projname <- paste0(modelname, "_", projs$scen)

df_probs <- pmap_dfr(projs, getprobs, df_proj_in)

saveRDS(df_probs, file = paste0(path, "df_probs.RDS"))
```

# Sum of probabilities and bins

```{r sum probabilities}
df_sums <- df_probs %>%
  dplyr::group_by(utmID, projname) %>%
  dplyr::summarise(probsum = sum(probs) / 1000,
                   binsum  = sum(bins))

df_idxy <- df_proj_in %>%
  filter(scen == "Current") %>%
  dplyr::select(utmID, X, Y)

df_sums <- left_join(df_sums, df_idxy, by = "utmID")

sumplots <- df_sums %>%
  split(.$projname) %>%
  map(., plotsum, path)
```

## Individual plots for each species (Current scenario only)

```{r species projection binary}
mypath <- paste0(path, "proj_bin/")
dir.create(mypath, showWarnings = FALSE)

scens <- unique(df_probs$projname)
scens <- scens[grepl("Current", x = scens)] # Comment this line to project all scenarios

for (scen in scens){
  plots <- df_probs %>%
    filter(projname == scen) %>%
    left_join(df_idxy, by = "utmID") %>%
    split(.$sp_n) %>%
    map(., plotbinaries, mypath, c("projname", "sp_n"))
}
```

```{r species projection continuous}
mypath <- paste0(path, "proj_prob/")
dir.create(mypath, showWarnings = FALSE)

for (scen in scens){
  plots <- df_probs %>%
    filter(projname == scen) %>%
    left_join(df_idxy, by = "utmID") %>%
    mutate(probs = probs / 1000) %>%
    split(.$sp_n) %>%
    map(., plotcontinuous, mypath, c("projname", "sp_n"), c(0, 1))
}
```
