==> R/09_NARA_S_Output.Rmd <==
---
title: "NARA_S_Output"
author: "Toon Van Daele"
date: "28 februari 2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
modelname <- "allsp"
path <- paste0("../data/data-out/", modelname, "/")
```

Generating output for NARA_S of model: **`r modelname`**

```{r loaddata}
df_probs <- readRDS(file = paste0(path, "df_probs.RDS"))
df_meval <- readRDS(file = paste0(path, "df_meval.RDS"))
df_redlist <- readRDS(file = paste0(path, "df_redlist.RDS"))

colnames(df_redlist)[1] <- "sp_n"
df_redlist$sp_n <- gsub("_", ".", df_redlist$sp_n)
df_redlist$sp_n <- gsub("-", ".", df_redlist$sp_n)
```

All species and all models

```{r all_species}
df_sums <- df_probs %>%
  dplyr::group_by(utmID, projname) %>%
  dplyr::summarise(probsum = sum(probs) / 1000) %>%
  spread(key = projname, value = probsum) %>%
  mutate_at(vars(contains("kr")),
                  .funs = funs(v = . - allsp_Current))

saveRDS(df_sums, file = paste0(path, "df_sums.RDS"))
write.table(df_sums, file = paste0(path, "df_sums.csv"),
            quote = FALSE, sep = ";", dec = ",", row.names = FALSE)
remove(df_sums)
```

Model selection: Select only species with ensemble AUC >= 0.7

```{r selection auc}
sp_sel <- df_meval %>%
  filter(Eval.metric == "ROC", median >= 0.7) %>%
  distinct(sp_n) %>%
  pull()

df_sums_auc <- df_probs %>%
  filter(sp_n %in% sp_sel) %>%
  dplyr::group_by(utmID, projname) %>%
  dplyr::summarise(probsum = sum(probs) / 1000) %>%
  spread(key = projname, value = probsum) %>%
  mutate_at(vars(contains("kr")),
                  .funs = funs(v = . - allsp_Current))

saveRDS(df_sums_auc, file = paste0(path, "df_sums_auc.RDS"))
write.table(df_sums_auc, file = paste0(path, "df_sums_auc.csv"),
            quote = FALSE, sep = ";", dec = ",", row.names = FALSE)
remove(df_sums_auc)
```

Weighted species according RL status

```{r weight RL status}
df_redlist <- df_redlist %>%
  select(sp_n, RLweging) %>%
  filter(!is.na(RLweging))

df_probs <- df_probs %>%
  left_join(df_redlist, by = "sp_n")

df_sums_rl <- df_probs %>%
  dplyr::group_by(utmID, projname) %>%
  dplyr::summarise(probsum = sum(probs * RLweging, na.rm = TRUE) / 1000) %>%
  spread(key = projname, value = probsum) %>%
  mutate_at(vars(contains("kr")),
                  .funs = funs(v = . - allsp_Current))

saveRDS(df_sums_rl, file = paste0(path, "df_sums_rl.RDS"))
write.table(df_sums_rl, file = paste0(path, "df_sums_rl.csv"),
            sep = ";", dec = ",")
remove(df_sums_rl)
```

Model selection: Select only species with ensemble AUC >= 0.7 & Weighted species according RL status

```{r weight RL status & auc}
df_sums_rl_auc <- df_probs %>%
  filter(sp_n %in% sp_sel) %>%
  dplyr::group_by(utmID, projname) %>%
  dplyr::summarise(probsum = sum(probs * RLweging, na.rm = TRUE) / 1000) %>%
  spread(key = projname, value = probsum) %>%
  mutate_at(vars(contains("kr")),
                  .funs = funs(v = . - allsp_Current))

saveRDS(df_sums_rl_auc, file = paste0(path, "df_sums_rl_auc.RDS"))
write.table(df_sums_rl_auc, file = paste0(path, "df_sums_rl_auc.csv"),
            sep = ";", dec = ",")
```
